+++ src/pages/text/splash_page.c
@@ -19,7 +19,7 @@
 
 static struct splash_obj * const gui = &gui_objs.u.splash;
 
-static u8 _action_cb(u32 button, u8 flags, void *data);
+static unsigned int _action_cb(u32 button, unsigned int flags, void *data);
 //static int offset =  110;
 
 void PAGE_SplashInit(int page)
@@ -37,7 +37,7 @@ void PAGE_SplashInit(int page)
     GUI_CreateLabelBox(&gui->version, 0, LCD_HEIGHT-1, LCD_WIDTH, ITEM_HEIGHT, &DEFAULT_FONT, NULL, NULL, DeviationVersion);
 }
 
-static u8 _action_cb(u32 button, u8 flags, void *data)
+static unsigned int _action_cb(u32 button, unsigned int flags, void *data)
 {
     (void)data;
     (void)button;
